Move db to separate class
