--- scoring_engine/database.py
@@ -1,18 +1,0 @@
-from sqlalchemy import create_engine
-from sqlalchemy.orm import scoped_session, sessionmaker
-from sqlalchemy.ext.declarative import declarative_base
-
-engine = create_engine('sqlite:////tmp/test.db',
-                       convert_unicode=True)
-db_session = scoped_session(sessionmaker(autocommit=False,
-                                         autoflush=False,
-                                        bind=engine))
-
-
-def init_db():
-    Model.metadata.create_all(bind=engine)
-
-
-Model = declarative_base(name='Model')
-Model.query = db_session.query_property()
-init_db()
