+++ scoring_engine/web/__init__.py
@@ -1,17 +1,14 @@
 import os
 
 from flask import Flask, g
-from flask_sqlalchemy import SQLAlchemy
-
 from flask_login import LoginManager, current_user
-
-from scoring_engine.web.views import welcome, scoreboard, overview, services, admin, auth
-
-from scoring_engine.models.user import User
-
 
 app = Flask(__name__)
 app.config.from_pyfile('settings.cfg')
+
+from scoring_engine.db import db_session
+
+from scoring_engine.web.views import welcome, scoreboard, overview, services, admin, auth
 
 app.register_blueprint(welcome.mod)
 app.register_blueprint(scoreboard.mod)
@@ -20,14 +17,14 @@ app.register_blueprint(services.mod)
 app.register_blueprint(admin.mod)
 app.register_blueprint(auth.mod)
 
-app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:////tmp/test.db'
 app.secret_key = os.urandom(128)
-db = SQLAlchemy(app)
-db.create_all()
 
 login_manager = LoginManager()
 login_manager.init_app(app)
 login_manager.login_view = 'login'
+
+
+from scoring_engine.models.user import User
 
 
 @login_manager.user_loader
