+++ scoring_engine/web/settings.cfg
@@ -1,2 +1,4 @@
 DEBUG = True
 SQLALCHEMY_TRACK_MODIFICATIONS = False
+DATABASE_URI = 'sqlite:////tmp/engine.db'
+DATABASE_CONNECT_OPTIONS = {}
